requirements.txt
fastapi
pydantic
mysql-connector-python
langchain
openai
uvicorn
